Move clock into own package
